pyproject.toml
[project]
name = "irai"
version = "0.1.0"
description = "IRAI — Offline-first voice assistant"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "openai-whisper>=20231117",
    "llama-cpp-python>=0.2.0",
    "piper-tts>=1.2.0",
    "sounddevice>=0.4.6",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "PyYAML>=6.0",
    "sentence-transformers>=2.2.0",
    "faiss-cpu>=1.7.4",
    "openwakeword>=0.5.0",
    "PyPDF2>=3.0.0",
]

[project.optional-dependencies]
gpu = [
    "faiss-gpu>=1.7.4",
]
dev = [
    "ruff>=0.1.0",
    "mypy>=1.0",
    "pytest>=7.0",
]

[project.scripts]
irai = "src.main:main"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"
